Update favicon and logo images, fix CORS origin, and handle file input click event
