+++ front/app/components/Footer.tsx
@@ -13,36 +13,6 @@ import NumberFormat from "react-number-format";
 export default function Footer() {
   const [currentRoute, setCurrentRoute] = useState("");
   const pathname = usePathname();
-  const [name, setName] = useState("");
-  const [email, setEmail] = useState("");
-  const [phone, setPhone] = useState("");
-  const [message, setMessage] = useState("");
-  const [formSubmitted, setFormSubmitted] = useState(false);
-  const [isSubmitting, setIsSubmitting] = useState(false);
-
-  const handleSubmit = async (event: { preventDefault: () => void }) => {
-    event.preventDefault();
-    setIsSubmitting(true);
-
-    const response = await fetch(
-      `${process.env.NEXT_PUBLIC_SUPERPET_DELIVERY_URL}:8080/contact`,
-      {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify({ name, email, phone, message }),
-      }
-    );
-
-    setIsSubmitting(false);
-
-    if (response.ok) {
-      setFormSubmitted(true);
-    } else {
-      console.error("Form submission failed");
-    }
-  };
 
   useEffect(() => {
     // Update the current route when the route changes
@@ -55,7 +25,7 @@ export default function Footer() {
         pathname === "/login" || pathname.startsWith("/admin") ? "hidden" : ""
       }`}>
       <Image src={FooterAnimals} alt='' className='wk-footer-image' />
-      <div className='container grid grid-cols-1 lg:grid-cols-4 p-20'>
+      <div className='container grid grid-cols-1 lg:grid-cols-3 p-20'>
         <div className='wk-footer-section'>
           <Image draggable={false} src={Logo} alt='Superpet' />
           <div className='wk-footer-text'>
@@ -98,68 +68,10 @@ export default function Footer() {
           </Link>
           <Link href='/produtos?category=Promoções'>Promoções</Link>
         </div>
-        <div className='wk-footer-section'>
+        {/* <div className='wk-footer-section'>
           <h3>Fale Conosco</h3>
-          <div className='wk-form'>
-            {formSubmitted ? (
-              <p>Obrigado pela mensagem, entraremos em contato em breve.</p>
-            ) : (
-              <form onSubmit={handleSubmit}>
-                <label>
-                  <h4>Nome</h4>
-                  <input
-                    type='text'
-                    name='name'
-                    required
-                    placeholder='Seu nome'
-                    value={name}
-                    onChange={(e) => setName(e.target.value)}
-                  />
-                </label>
-                <label>
-                  <h4>Email</h4>
-                  <input
-                    type='email'
-                    name='email'
-                    required
-                    placeholder='Seu email'
-                    value={email}
-                    onChange={(e) => setEmail(e.target.value)}
-                  />
-                </label>
-                <label>
-                  <h4>Telefone</h4>
-                  <NumberFormat
-                    format='(##) # ####-####'
-                    isNumericString={true}
-                    value={phone}
-                    onValueChange={(values: {
-                      value: SetStateAction<string>;
-                    }) => setPhone(values.value)}
-                    placeholder='Ex: (00) 9 9999-9999'
-                  />
-                </label>
-                <label>
-                  <h4>Mensagem</h4>
-                  <textarea
-                    name='message'
-                    cols={30}
-                    rows={10}
-                    required
-                    placeholder='Sua mensagem'
-                    value={message}
-                    onChange={(e) => setMessage(e.target.value)}></textarea>
-                </label>
-                <button
-                  className='wk-btn wk-btn--sm wk-btn--primary'
-                  type='submit'
-                  disabled={isSubmitting}>
-                  {isSubmitting ? "Enviando..." : "Enviar"}
-                </button>
-              </form>
-            )}
-          </div>
-        </div>
+         
+        </div> */}
       </div>
       <div className='wk-footer-main container p-20 flex items-center justify-end'>
         <WKode />
@@ -170,7 +82,9 @@ export default function Footer() {
         <Image
           src={WhatsappImage}
           alt='Imagem whatsapp'
-          className='wk-whatsapp-image'
+          className={`wk-whatsapp-image ${
+            pathname === "/contato" ? "wk-whatsapp-image--footer" : ""
+          }`}
           width={70}
           height={70}
         />
