+++ front/app/styles/admin/header.scss
@@ -15,7 +15,7 @@
   width: 100%;
   z-index: 999;
   &__logo {
-    left: 0.3rem;
+    left: 1rem;
     position: absolute;
     top: 0.4rem;
   }
