+++ front/package.json
@@ -28,6 +28,7 @@
     "react-beautiful-dnd": "^13.1.1",
     "react-dom": "^18",
     "react-google-recaptcha": "^3.1.0",
+    "react-google-recaptcha-v3": "^1.10.1",
     "react-icons": "^4.12.0",
     "react-number-format": "^4.9.4",
     "react-slick": "^0.30.2",
