+++ front/public/static/images/favicon.ico
